기업과제3/FastAPI/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64

fastapi=0.70.1=pyhd8ed1ab_0
gdown=4.4.0=pyhd8ed1ab_0
pydantic=1.9.0=py37hcc03f2d_0
pytorch=1.11.0=py3.7_cuda11.3_cudnn8_0
pytorch-mutex=1.0=cuda
sentence-transformers=2.2.0=pypi_0
torchaudio=0.11.0=py37_cu113
torchvision=0.12.0=py37_cu113
uvicorn=0.17.6=py37h03978a9_0
